+ inference, training, model correctness fixes (val_dataloader guard, HGT CPU attention, Bayesian sampling, checkpoint resume, AMP softmax, masked softmax, config validation)

+++ src/models/fusion/pathology_attention.py
@@ -175,7 +175,7 @@ class PathologyStratifiedAttention(nn.Module):
 
         # Softmax and attend
         assert scores.size(2) == 1, f"Expected query dim 1, got {scores.size(2)}"
-        attention_weights = F.softmax(scores, dim=-1)[:, :, 0, :]  # [B, n_heads, n_cell_types]
+        attention_weights = F.softmax(scores.float(), dim=-1).to(scores.dtype)[:, :, 0, :]  # [B, n_heads, n_cell_types]
 
         # Zero out attention for fully-masked samples (softmax gave near-uniform,
         # but these samples should contribute nothing)
